+++ sdk/src/driftClient.ts
@@ -5888,13 +5888,15 @@ export class DriftClient {
 		return this.signMessage(takerOrderParamsMessage);
 	}
 
+	// encode the swift order for use in program Ix/signing
 	public encodeSwiftOrderParamsMessage(
 		orderParamsMessage: SwiftOrderParamsMessage
 	): Buffer {
-		return this.program.coder.types.encode(
+		const borshBuf = this.program.coder.types.encode(
 			'SwiftOrderParamsMessage',
 			orderParamsMessage
 		);
+		return Buffer.from(borshBuf.toString('hex'));
 	}
 
 	public decodeSwiftOrderParamsMessage(
@@ -9200,6 +9202,35 @@ export class DriftClient {
 		);
 
 		return ix;
+	}
+
+	public async getPauseSpotMarketDepositWithdrawIx(
+		spotMarketIndex: number
+	): Promise<TransactionInstruction> {
+		const spotMarket = await this.getSpotMarketAccount(spotMarketIndex);
+		return this.program.instruction.pauseSpotMarketDepositWithdraw({
+			accounts: {
+				state: await this.getStatePublicKey(),
+				keeper: this.wallet.publicKey,
+				spotMarket: spotMarket.pubkey,
+				spotMarketVault: spotMarket.vault,
+			},
+		});
+	}
+
+	public async pauseSpotMarketDepositWithdraw(
+		spotMarketIndex: number,
+		txParams?: TxParams
+	): Promise<TransactionSignature> {
+		const { txSig } = await this.sendTransaction(
+			await this.buildTransaction(
+				await this.getPauseSpotMarketDepositWithdrawIx(spotMarketIndex),
+				txParams
+			),
+			[],
+			this.opts
+		);
+		return txSig;
 	}
 
 	private handleSignedTransaction(signedTxs: SignedTxData[]) {