+++ CHANGELOG.md
@@ -11,6 +11,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixes
 
+- program: account for fuel in swaps ([#1411](https://github.com/drift-labs/protocol-v2/pull/1411))
+- program: account for fuel when there is full withdraw ([#1413](https://github.com/drift-labs/protocol-v2/pull/1413))
+
+### Breaking
+
+## [2.105.0] - 2025-01-02
+
+### Features
+
+- program: add ix to pause deposits/withdraws if vault invariant broken ([#1387](https://github.com/drift-labs/protocol-v2/pull/1387))
+
+### Fixes
+
+- program: fix spot swap fuel bonus ([#1411](https://github.com/drift-labs/protocol-v2/pull/1411))
+- program: skip liq perp oracle twap check if market is in settlement ([#1406](https://github.com/drift-labs/protocol-v2/pull/1406))
+
 ### Breaking
 
 ## [2.104.0] - 2024-12-23